Python and Julia wrapper tidy
Python:

Two files used during the setup of the Python wrapper, agdeblend_libs.py
and agdeblend_build.py, were previously inside the agdeblend package.
This caused a problem as it required importing the package before
the dependencies had been installed. These files have now been moved
out of the package.

A new check was also added to only register the MPI libraries as
having been found when installing the Python wrapper if mpi4py
is installed. The setup could otherwise fail if it tries to
finds MPI libraries and so tries to import mpi4py.

Julia:

The MPI package was always imported, so would fail if it was not
installed. A new "FakeMPI" module is used instead if it is not
installed.