Fix wavelet memory access bug
When wavelets are being used the memory allocated for the
temporary arrays temp_1 and temp_2 may not have been sufficient.
This is resolved by adding the number of bytes that will be
needed when Fourier transforming the data for wavelet correlation
to the list of memory requirements that are considered when
allocating these arrays.